totalPrice implementation

+++ src/main/java/com/booleanuk/core/Coffee.java
@@ -1,8 +1,10 @@
 package com.booleanuk.core;
 
+import java.math.BigDecimal;
+
 public class Coffee extends Item{
 
-    public Coffee(String SKU, String name, Double price, String variant){
+    public Coffee(String SKU, String name, BigDecimal price, String variant){
         super(SKU,name,price,variant);
     }
